feat: add resource allocation interaction (1st step) (#10)
When clicking on the "Allocate more resource" button, the sub-process
diagram is displayed.
- It's running task is highlighted and the path leading to it is
modified as well. This task proposed a popover which opens on hover.
- In the future, the style of some elements of the diagram will be
modified when hovering on the row of the popover.
When clicking on the "Contact-Client" button, a window alert is
displayed. In the future, it will display the 2nd process of the diagram
and it will let interact with it.


### Implementation notes

This PR introduces refactoring to better encapsulate the logic.
In particular, it provides classes that may be reused outside this demo
and served as starting point for experimental features
- BpmnElementsSearcher
- PathResolver

The case monitoring data are fully managed by dedicated classes both for
the main process and the sub-process.
Other parts of the code are still leaking the use cases at several
places. This will be refactored later to limit the size of this PR.

---------

Co-authored-by: Thomas Bouffard <27200110+tbouffard@users.noreply.github.com>