refactor: do lazy initializations and improve view management
Treat the initialization and loading of the main diagram in the same way as we do for the subprocess
diagram.
This involves lazily loading the initialization of the various parts that previously assumed that
the main diagram was loaded and thus, that searches always returned elements

The other refactoring of the views mainly concerns the improvement of the naming of functions and
variables.