fix(case-monitoring): remove listeners on popover elements (#24)
In the case monitoring, the listeners on the elements within the
popovers were not removed. There were side effects when the popover show
again as there were several listeners to do the same thing.
When quitting the sub-process and going back, the popover didn't
highlight the activities. This is now fixed by using a more precise CSS
selector.

Additional fix: when displaying the sub-process of the case-monitoring
and going to the "none" use case, the sub-process still contained the
highlight element and its related popover