feat: introduce `ProcessExecutor` to manage Supplier Contact execution
Manage the simulation of the execution of the "Supplier Contact" process execution.
It lets configure the execution and the highlight of the paths.
In the future, interaction will also be managed.

Also start improving the global asynchronous execution (this is still a work in progress, a lot
of things remains to do).