feat: introduce `ProcessExecutor` for the Supplier Contact process (#32)
Manage the simulation of the execution of the "Supplier Contact" process
execution. It lets configure the execution and the highlight of the
paths. In the future, interactions with the "Review and adjust email"
popover will also be managed.
The interactions are currently simulated by clicking on the elements of
the diagram
- "review and adjust email": request new suggestion
- gateway: send email
- abort end event: abort

Also start improving the global asynchronous execution (this is still a
work in progress, a lot of things remains to do).