refactor: reorganize notifications (#95)
Introduce a class to hide the 'Notyf' library to the code using the
notification. This makes the notification easier to create and remove
adherence to an external library at several places in the application.

Don't hard code the condition that triggers the display of notifications
in `ProcessExecutor`. `ProcessExecutor` is intended to be as general as
possible for a future near reuse outside of this demonstration
Notifications are now made configurable in the `ExecutionStep`.

Change NotificationType from enum to type. There is no need to enumerate
the values and TS enum doesn't operate very well with JS code, so let's
avoid them when alternatives exist.