refactor(case-monitoring): simplify sub-process hide data (#102)
This is possible thanks to a fix available in bpmn-visualization 0.33.1.
Previously, the data could only be hidden if the sub-process diagram was
already loaded. This forced to leak the internal state and made the code
harder to understand.

Also improve the robustness of the PathResolver to handle shapeId that
are not available in the diagram. Here, this lets support not loaded
diagram.